labs/lab05/app/requirements.txt
flask==3.1.*
opentelemetry-api==1.29.*
opentelemetry-sdk==1.29.*
opentelemetry-exporter-otlp==1.29.*
opentelemetry-instrumentation-flask==0.50b.*
opentelemetry-distro==0.50b.*
opentelemetry-instrumentation==0.50b.*
